backend/requirements.txt
Django==3.1.4
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-cors-headers==3.6.0
djangorestframework-jwt==1.11.0